requirements.txt
requests==2.22.0
python-telegram-bot==20.2
Pillow==9.3.0
